Revert "Ed 09 06 2024 upstream (#230)"
This reverts commit 50470c3aaa28be2319951b05d80353423536455f.

+++ Content.Server/Pointing/EntitySystems/PointingSystem.cs
@@ -174,7 +174,7 @@ namespace Content.Server.Pointing.EntitySystems
             {
                 var arrowVisibility = EntityManager.EnsureComponent<VisibilityComponent>(arrow);
                 layer = playerVisibility.Layer;
-                _visibilitySystem.SetLayer((arrow, arrowVisibility), (ushort) layer);
+                _visibilitySystem.SetLayer(arrow, arrowVisibility, layer);
             }
 
             // Get players that are in range and whose visibility layer matches the arrow's.
